chore: DOM 로직을 domain으로부터 분리
