+++ src/components/SortedRestaurant.ts
@@ -1,5 +1,3 @@
-import { SortByDistance, SortByName } from "../domain/SortRestaurantByValue";
-
 const SortType = {
   NAME: "name",
   DISTANCE: "distance",
@@ -27,5 +25,32 @@ export default function SortedRestaurant() {
       sortedRestaurants = SortByDistance($restaurantItems);
     }
     $restaurantList.replaceChildren(...sortedRestaurants);
+  });
+}
+
+function SortByName($restaurantItems: HTMLElement[]) {
+  return $restaurantItems.sort((a, b) => {
+    const aName =
+      (a.querySelector(".restaurant__name") as HTMLElement)?.textContent ?? "";
+    const bName =
+      (b.querySelector(".restaurant__name") as HTMLElement)?.textContent ?? "";
+
+    return aName.localeCompare(bName);
+  });
+}
+
+function SortByDistance($restaurantItems: HTMLElement[]) {
+  return $restaurantItems.sort((a, b) => {
+    const aDistanceText =
+      (a.querySelector(".restaurant__distance") as HTMLElement)?.textContent ??
+      "";
+    const bDistanceText =
+      (b.querySelector(".restaurant__distance") as HTMLElement)?.textContent ??
+      "";
+
+    const aDistance = parseFloat(aDistanceText.replace(/[^0-9]/g, "")) || 0;
+    const bDistance = parseFloat(bDistanceText.replace(/[^0-9]/g, "")) || 0;
+
+    return aDistance - bDistance;
   });
 }
